Update download and decompress methods
